feature dns13 update 2.1
