--- src/app/index.tsx
@@ -1,12 +1,0 @@
-import {Link} from "react-router-dom"
-const App = () => {
-
-  return (
-    <>
-      <h1>hello world</h1>
-      <Link to="myname">fdsf</Link>
-    </>
-  )
-}
-
-export default App
