+++ src/pages/main/index.tsx
@@ -1,8 +1,12 @@
 import React from 'react'
+import { Link } from 'react-router-dom'
 
 const Main: React.FC = () => {
     return (
-      <div>Main</div>
+      <div>
+        <h1>hello 1world</h1>
+      <Link to="career">fdsf</Link>
+      </div>
     )
 }
